Replace kotlinx.coroutines.reactive.FlowKt and kotlinx.coroutines.reactor.FlowKt with properly named files
Otherwise, it leads to a weird classname clash and forces to write FQ imports
