+++ reactive/kotlinx-coroutines-reactor/src/ReactorFlow.kt
@@ -1,4 +1,6 @@
-@file:JvmName("FlowKt")
+/*
+ * Copyright 2016-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
+ */
 
 package kotlinx.coroutines.reactor
